Integrate stats page into course monitor
